requirements.txt
# GUI 依赖 - PyQt5 (桌面版)
PyQt5>=5.15.0

# Web GUI 依赖 - NiceGUI (可选)
nicegui>=1.4.0

# 加密库依赖
pycryptodome>=3.15.0
cryptography>=3.4.0
